Rewrited GUICreatorAPI, New methods in ItemCreatorAPI, Fixed TabCompleter for /TheAPI command, Fixed %math{...}% placeholder, BossBars for 1.8+

+++ src/me/DevTec/NumbersAPI.java
@@ -18,10 +18,10 @@ public class NumbersAPI {
 
 	/**
 	 * @see see Calculate string
-	 * @return double
+	 * @return float
 	 */
-	public double calculate() {
-		return TheAPI.getStringUtils().calculate(fromString);
+	public float calculate() {
+		return TheAPI.getStringUtils().calculate(fromString).floatValue();
 	}
 
 	/**
